almost working init
